backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.5.2
python-dotenv==1.0.0
motor==3.3.2
python-multipart==0.0.6
email-validator==2.3.0
PyJWT==2.9.0
passlib>=1.7.4
bcrypt>=4.0.0
httpx==0.24.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
anyio>=3.7.1,<4.0.0